+++ include/flatbuffers/util.h
@@ -22,7 +22,6 @@
 
 #include "flatbuffers/base.h"
 #include "flatbuffers/stl_emulation.h"
-#include "flatbuffers/dragonbox_to_chars.h"
 
 #ifndef FLATBUFFERS_PREFER_PRINTF
 #include <iomanip>
@@ -156,42 +155,9 @@ inline std::string NumToString<char>(char t) {
   return NumToString(static_cast<int>(t));
 }
 
+// definitions in util.cpp
 template <typename Float>
-std::string format_fixed_dragonbox(Float value) {
-    auto dec = jkj::dragonbox::to_decimal(value);
-
-    std::string digits = std::to_string(dec.significand);
-    const int k = static_cast<int>(digits.size());
-    const int exp = dec.exponent;
-
-    std::string out;
-
-    if (dec.is_negative) {
-        out.push_back('-');
-    }
-
-    if (exp >= 0) {
-        out += digits;
-        out.append(exp, '0');
-    }
-    else if (exp > -k) {
-        const int pos = k + exp;
-        out.append(digits, 0, pos);
-        out.push_back('.');
-        out.append(digits, pos, std::string::npos);
-    }
-    else {
-        out += "0.";
-        out.append(-exp - k, '0');
-        out += digits;
-    }
-
-    if (out.find('.') == std::string::npos) {
-        out += ".0";
-    }
-
-    return out;
-}
+std::string format_fixed_dragonbox(Float value);
 
 // Special versions for floats/doubles.
 template <typename T>
@@ -199,21 +165,6 @@ std::string FloatToString(T t, [[maybe_unused]] int precision) {
   // clang-format off
 
   #ifndef FLATBUFFERS_PREFER_PRINTF
-    // Handle NaN
-    if (std::isnan(t)) {
-        return "nan";
-    }
-
-    // Handle infinities
-    if (std::isinf(t)) {
-        return std::signbit(t) ? "-inf" : "inf";
-    }
-
-    // Handle zero (preserve sign of -0.0 if desired)
-    if (t == T{0}) {
-        return "0.0";
-    }
-
     return format_fixed_dragonbox(t);
   #else // FLATBUFFERS_PREFER_PRINTF
     auto v = static_cast<double>(t);
